Split statistics table to separate component
This makes it easier to add other things such as graphs.
